backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LLM Providers (Hybrid)
openai>=1.12.0
ollama>=0.1.6

# Embeddings & Vector Search
sentence-transformers>=2.2.0
numpy>=1.24.0

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
